Changed backend to work with doubles

See merge request !26

+++ src/backend/turtlecommands/PositionMove.java
@@ -10,7 +10,7 @@ public abstract class PositionMove extends Command {
 	}
 	
 	public String setPosition(int x, int y, Turtle t) {
-		Integer distance = computeDistance(x,y,t.getMyX(),t.getMyY());
+		Double distance = computeDistance(x,y,t.getMyX(),t.getMyY());
 		
 		//should this draw a line?
 		t.setMyX(x);
@@ -20,9 +20,9 @@ public abstract class PositionMove extends Command {
 		
 	}
 	
-	private Integer computeDistance(int x1, int y1, int x2, int y2) {
+	private Double computeDistance(double x1, double y1, double x2, double y2) {
 		
-		return (int) Math.sqrt( Math.pow(x1-x2, 2) + Math.pow(y1-y2, 2) );
+		return Math.sqrt( Math.pow(x1-x2, 2) + Math.pow(y1-y2, 2) );
 		
 	}
